Apply Phase 2 pooling GTest migration on top of develop
