+++ projects/miopen/test/gtest/pooling2d_wide_nhwc.cpp
@@ -28,141 +28,66 @@
 #include <miopen/env.hpp>
 #include "get_handle.hpp"
 #include "gtest_common.hpp"
-#include "pooling2d.hpp"
-
-MIOPEN_DECLARE_ENV_VAR_STR(MIOPEN_TEST_FLAGS_ARGS)
-
-namespace env = miopen::env;
+#include "pooling2d_common.hpp"
 
 namespace pooling2d_wide_nhwc {
 
-class GPU_WidePooling2d_NHWC_FP32 : public testing::TestWithParam<std::vector<std::string>>
-{
-    MIOPEN_DECLARE_GTEST_USES_TEST_DRIVE();
-};
+using namespace pooling2d_gtest;
 
-class GPU_WidePooling2d_NHWC_FP16 : public testing::TestWithParam<std::vector<std::string>>
-{
-    MIOPEN_DECLARE_GTEST_USES_TEST_DRIVE();
-};
+struct GPU_WidePooling2d_NHWC_FP32 : public Pooling2dCommon<float> {};
+struct GPU_WidePooling2d_NHWC_FP16 : public Pooling2dCommon<half_float::half> {};
+struct GPU_WidePooling2d_NHWC_BFP16 : public Pooling2dCommon<bfloat16> {};
 
-class GPU_WidePooling2d_NHWC_BFP16 : public testing::TestWithParam<std::vector<std::string>>
+std::vector<PoolingTestCase> GetTestCases(miopenDataType_t prec)
 {
-    MIOPEN_DECLARE_GTEST_USES_TEST_DRIVE();
-};
+    std::vector<int> wsidx_values = {0, 1};
+    std::vector<miopenPoolingMode_t> modes = {miopenPoolingMax, miopenPoolingAverage, miopenPoolingAverageInclusive};
+    std::vector<miopenIndexType_t> index_types = {miopenIndexUint8, miopenIndexUint16, miopenIndexUint32, miopenIndexUint64};
 
-void GetArgs(const std::string& param, std::vector<std::string>& tokens)
-{
-    std::stringstream ss(param);
-    std::istream_iterator<std::string> begin(ss);
-    std::istream_iterator<std::string> end;
-    while(begin != end)
-        tokens.push_back(*begin++);
-}
+    // Dataset 2: Wide window configs
+    std::vector<std::vector<int>> in_shapes = {{1, 3, 255, 255}, {2, 3, 227, 227}, {1, 7, 127, 127}, {1, 1, 410, 400}};
+    std::vector<std::vector<int>> lens_list = {{35, 35}, {100, 100}, {255, 255}, {410, 400}};
+    std::vector<std::vector<int>> strides_list = {{1, 1}};
+    std::vector<std::vector<int>> pads_list = {{0, 0}};
 
-void Run2dDriver(miopenDataType_t prec)
-{
-
-    std::vector<std::string> params;
-    switch(prec)
+    std::vector<PoolingTestCase> test_cases;
+    for(const auto& in_shape : in_shapes)
     {
-    case miopenFloat: params = GPU_WidePooling2d_NHWC_FP32::GetParam(); break;
-    case miopenHalf: params = GPU_WidePooling2d_NHWC_FP16::GetParam(); break;
-    case miopenBFloat16: params = GPU_WidePooling2d_NHWC_BFP16::GetParam(); break;
-    case miopenInt8:
-    case miopenFloat8_fnuz:
-    case miopenBFloat8_fnuz:
-    case miopenInt32:
-    case miopenInt64:
-    case miopenDouble:
-        FAIL() << "miopenInt8, miopenInt32, miopenDouble, miopenFloat8_fnuz, "
-                  "miopenBFloat8_fnuz "
-                  "data type not supported by "
-                  "pooling2d_wide_nhwc test";
-
-    default: params = GPU_WidePooling2d_NHWC_FP32::GetParam();
+        AddTestCasesForInput(in_shape,
+                             lens_list,
+                             strides_list,
+                             pads_list,
+                             index_types,
+                             modes,
+                             wsidx_values,
+                             test_cases,
+                             false, // skip_wide_check
+                             true,  // is_wide_dataset
+                             "NHWC",
+                             "NHWC");
     }
-
-    for(const auto& test_value : params)
-    {
-        std::vector<std::string> tokens;
-        GetArgs(test_value, tokens);
-        std::vector<const char*> ptrs;
-
-        std::transform(tokens.begin(), tokens.end(), std::back_inserter(ptrs), [](const auto& str) {
-            return str.data();
-        });
-
-        testing::internal::CaptureStderr();
-        test_drive<pooling2d_driver>(ptrs.size(), ptrs.data());
-        auto capture = testing::internal::GetCapturedStderr();
-        std::cerr << capture;
-    }
-};
-
-bool IsTestSupportedForDevice() { return true; }
-
-std::vector<std::string> GetTestCases(const std::string& precision)
-{
-    const auto& flag_arg = env::value(MIOPEN_TEST_FLAGS_ARGS);
-
-    return std::vector<std::string>{
-        // clang-format off
-        // Forward pooling with NHWC layout (wide windows, batched transpose)
-        {"test_pooling2d " + precision + " --all --dataset 2 --limit 0 --in_layout NHWC --out_layout NHWC " + flag_arg},
-        // Backward pooling with NHWC layout (batched transpose)
-        {"test_pooling2d " + precision + " --forw 0 --in_layout NHWC --out_layout NHWC " + flag_arg}
-        // clang-format on
-    };
+    return test_cases;
 }
 
 } // namespace pooling2d_wide_nhwc
+
 using namespace pooling2d_wide_nhwc;
 
-TEST_P(GPU_WidePooling2d_NHWC_FP32, FloatTest_pooling2d_wide_nhwc)
-{
-    if(IsTestSupportedForDevice())
-    {
-        Run2dDriver(miopenFloat);
-    }
-    else
-    {
-        GTEST_SKIP();
-    }
-};
-
-TEST_P(GPU_WidePooling2d_NHWC_FP16, HalfTest_pooling2d_wide_nhwc)
-{
-    if(IsTestSupportedForDevice())
-    {
-        Run2dDriver(miopenHalf);
-    }
-    else
-    {
-        GTEST_SKIP();
-    }
-};
-
-TEST_P(GPU_WidePooling2d_NHWC_BFP16, BFloat16Test_pooling2d_wide_nhwc)
-{
-    if(IsTestSupportedForDevice())
-    {
-        Run2dDriver(miopenBFloat16);
-    }
-    else
-    {
-        GTEST_SKIP();
-    }
-};
+TEST_P(GPU_WidePooling2d_NHWC_FP32, FloatTest) { RunTest(); }
+TEST_P(GPU_WidePooling2d_NHWC_FP16, HalfTest) { RunTest(); }
+TEST_P(GPU_WidePooling2d_NHWC_BFP16, BFloat16Test) { RunTest(); }
 
 INSTANTIATE_TEST_SUITE_P(Full,
                          GPU_WidePooling2d_NHWC_FP32,
-                         testing::Values(GetTestCases("--float")));
+                         testing::ValuesIn(GetTestCases(miopenFloat)),
+                         GetPoolingTestCaseName);
 
 INSTANTIATE_TEST_SUITE_P(Full,
                          GPU_WidePooling2d_NHWC_FP16,
-                         testing::Values(GetTestCases("--half")));
+                         testing::ValuesIn(GetTestCases(miopenHalf)),
+                         GetPoolingTestCaseName);
 
 INSTANTIATE_TEST_SUITE_P(Full,
                          GPU_WidePooling2d_NHWC_BFP16,
-                         testing::Values(GetTestCases("--bfloat16")));
+                         testing::ValuesIn(GetTestCases(miopenBFloat16)),
+                         GetPoolingTestCaseName);
