+++ projects/miopen/test/gtest/pooling_gtest_common.hpp
@@ -7,18 +7,24 @@
 #include <algorithm>
 #include <array>
 #include <cstddef>
+#include <cmath>
 #include <iostream>
 #include <limits>
 #include <numeric>
+#include <string>
+#include <unordered_map>
 #include <miopen/logger.hpp>
 #include <miopen/miopen.h>
 #include <miopen/pooling.hpp>
+#include <miopen/stringutils.hpp>
 #include <miopen/tensor.hpp>
+#include <numeric>
 #include <miopen/tensor_layout.hpp>
 #include <utility>
 
 // #include "../network_data.hpp"
 #include "../cpu_conv.hpp"
+#include "../driver.hpp"
 #include "../tensor_holder.hpp"
 #include "../test.hpp"
 #include "../verify.hpp"
@@ -26,6 +32,16 @@
 #include "get_handle.hpp"
 
 #define TEST_PADDING_MODE 0
+// NOLINTNEXTLINE (cppcoreguidelines-avoid-non-const-global-variables)
+static int num_uint16_case = 0;
+// NOLINTNEXTLINE (cppcoreguidelines-avoid-non-const-global-variables)
+static int num_uint32_case = 0;
+// NOLINTNEXTLINE (cppcoreguidelines-avoid-non-const-global-variables)
+static int num_uint32_case_imgidx = 0;
+// NOLINTNEXTLINE (cppcoreguidelines-avoid-non-const-global-variables)
+static int num_uint64_case = 0;
+// NOLINTNEXTLINE (cppcoreguidelines-avoid-non-const-global-variables)
+static int num_uint64_case_imgidx = 0;
 
 static inline void print(const miopen::PoolingDescriptor& filter)
 {
@@ -494,6 +510,397 @@ struct verify_backward_pooling
         print(filter);
         std::cout << "Input tensor: " << input.desc.ToString() << std::endl;
         std::cout << "Output tensor: " << out.desc.ToString() << std::endl;
+    }
+};
+
+template <class T>
+struct pooling_harness : test_driver
+{
+    miopen::PoolingDescriptor filter;
+    std::vector<int> in_shape;
+    std::vector<int> lens;
+    std::vector<int> pads;
+    std::vector<int> strides;
+    std::string index_type;
+    std::string mode;
+#if TEST_PADDING_MODE == 1
+    std::string pmode;
+#endif
+    int verify_indices{};
+    int wsidx{};
+    std::string in_layout;
+    std::unordered_map<std::string, miopenIndexType_t> index_type_lookup = {
+        {miopen::ToUpper("miopenIndexUint8"), miopenIndexUint8},
+        {miopen::ToUpper("miopenIndexUint16"), miopenIndexUint16},
+        {miopen::ToUpper("miopenIndexUint32"), miopenIndexUint32},
+        {miopen::ToUpper("miopenIndexUint64"), miopenIndexUint64},
+    };
+    std::unordered_map<std::string, miopenPoolingMode_t> mode_lookup = {
+        {"MAX", miopenPoolingMax},
+        {"MIOPENPOOLINGMAX", miopenPoolingMax},
+        {"AVERAGE", miopenPoolingAverage},
+        {"MIOPENPOOLINGAVERAGE", miopenPoolingAverage},
+        {"AVERAGEINCLUSIVE", miopenPoolingAverageInclusive},
+        {"MIOPENPOOLINGAVERAGEINCLUSIVE", miopenPoolingAverageInclusive},
+    };
+#if TEST_PADDING_MODE == 1
+    std::unordered_map<std::string, miopenPaddingMode_t> pmode_lookup = {
+        {"DEFAULT", miopenPaddingDefault},
+        {"SAME", miopenPaddingSame},
+        {"VALID", miopenPaddingValid},
+    };
+#endif
+    pooling_harness()
+    {
+        add(index_type,
+            "index_type",
+            generate_multi_data<const char*>( //
+                {{"miopenIndexUint8",
+                  "miopenIndexUint16",
+                  "miopenIndexUint32",
+                  "miopenIndexUint64"},                     //
+                 {"miopenIndexUint8", "miopenIndexUint32"}, //
+                 {"miopenIndexUint32"}}                     //
+                ));
+        add(mode,
+            "mode",
+            generate_data(
+                {"miopenPoolingMax", "miopenPoolingAverage", "miopenPoolingAverageInclusive"}));
+#if TEST_PADDING_MODE == 1
+        add(pmode, "pmode", generate_data({"default", "same", "valid"}));
+#endif
+        add(verify_indices, "verify_indices", generate_data({1}));
+        add(in_layout, "in_layout", generate_data({"NCHW"}));
+    }
+
+    template <class Index, int SptDim>
+    void run_impl()
+    {
+        std::vector<Index> indices{};
+        auto input = tensor<T>{in_shape}.generate(tensor_elem_gen_integer{
+            (miopen_type<T>{} == miopenHalf || miopen_type<T>{} == miopenBFloat16) ? 5 : 17});
+        if(in_layout != "NCHW" && in_layout != "NCDHW")
+        {
+            const std::vector<std::size_t> dim_lens = input.desc.GetLengths();
+            std::vector<std::size_t> dim_strides;
+            miopen::tensor_layout_to_strides(
+                dim_lens,
+                miopen::tensor_layout_get_default(input.desc.GetNumDims()),
+                in_layout,
+                dim_strides);
+            input.desc = miopen::TensorDescriptor(miopen_type<T>{}, dim_lens, dim_strides);
+        }
+        auto out  = verify(verify_forward_pooling<SptDim>{}, input, filter, indices);
+        auto dout = out.first;
+        dout.generate(tensor_elem_gen_integer{2503});
+        verify(verify_backward_pooling<SptDim>{},
+               input,
+               dout,
+               out.first,
+               filter,
+               indices,
+               wsidx != 0,
+               static_cast<bool>(this->verify_indices));
+    }
+
+    void run()
+    {
+        auto idx_typ = index_type_lookup.at(miopen::ToUpper(index_type));
+        auto idx_sz  = sizeof(uint8_t);
+        int spt_dim  = in_shape.size() - 2;
+        const bool skip_many_configs_with_non_int8_index =
+            (dataset_id == 0) && full_set; // Otherwise the default dataset takes too much time.
+        const bool wide_dataset = (dataset_id == 2) && full_set;
+
+        filter = miopen::PoolingDescriptor
+        {
+            mode_lookup.at(miopen::ToUpper(mode)),
+#if TEST_PADDING_MODE == 1
+                pmode_lookup.at(miopen::ToUpper(pmode)),
+#else
+                miopenPaddingDefault,
+#endif
+                lens, strides, pads
+        };
+
+        filter.SetIndexType(idx_typ);
+        filter.SetWorkspaceIndexMode(miopenPoolingWorkspaceIndexMode_t(wsidx));
+
+        if(wsidx == 0 && spt_dim == 3 && filter.GetMode() == miopenPoolingMax && full_set)
+        {
+            show_command();
+            std::cout << "Warning: Config skipped. Workspace index mask mode is not implemented "
+                         "yet in 3D max pooling solvers."
+                      << std::endl;
+            return;
+        }
+
+        if(wsidx == 0 && spt_dim == 2 && filter.GetMode() == miopenPoolingMax && wide_dataset)
+        {
+            show_command();
+            std::cout << "Warning: Config skipped. Workspace index mask mode is not implemented "
+                         "yet in 2D max backward solvers that support wide pooling window."
+                      << std::endl;
+            return;
+        }
+
+        if(wsidx == 0 &&
+           (filter.GetMode() == miopenPoolingAverage ||
+            filter.GetMode() == miopenPoolingAverageInclusive) &&
+           full_set)
+        {
+            show_command();
+            std::cout << "Warning: Config skipped. Workspace index modes are irrelevant for "
+                         "Average pooling. "
+                         "In order to optimize performance of full tests, we "
+                         "skip average pooling configs when (wsidx == 0). "
+                         "Please make sure that dataset includes counterparts with (wsidx == 1)."
+                      << std::endl;
+            return;
+        }
+
+        switch(idx_typ)
+        {
+        /// The "index is too small" limitation is an approximation
+        /// of the real limitation, and therefore applied only when
+        /// the "full test" is ran. See:
+        /// \ref max_pooling_index_max_restriction
+        case miopenIndexUint8: {
+            if((spt_dim == 3 || (spt_dim == 2 && wsidx == 1)) && full_set &&
+               filter.GetMode() == miopenPoolingMax)
+            {
+                show_command();
+                std::cout << "Warning: Config skipped: uint8 index is too small "
+                             "(spt_dim == 3 || (spt_dim == 2 && wsidx == 1)) "
+                             "&& filter.GetMode() == miopenPoolingMax"
+                          << std::endl;
+                return;
+            }
+            break;
+        }
+        case miopenIndexUint16: {
+            if((spt_dim == 3 || (spt_dim == 2 && wsidx == 1)) && full_set &&
+               filter.GetMode() == miopenPoolingMax)
+            {
+                show_command();
+                std::cout << "Warning: Config skipped: uint16 index is too small "
+                             "(spt_dim == 3 || (spt_dim == 2 && wsidx == 1)) "
+                             "&& filter.GetMode() == miopenPoolingMax"
+                          << std::endl;
+                return;
+            }
+            if(skip_many_configs_with_non_int8_index)
+            {
+                // test_pooling_test --all only test 5 uint16 cases
+                if(num_uint16_case > 5)
+                {
+                    show_command();
+                    std::cout << "Warning: Config skipped for the default dataset to speed "
+                                 "up testing (num_uint16_case > 5)"
+                              << std::endl;
+                    return;
+                }
+                ++num_uint16_case;
+            }
+            idx_sz = sizeof(uint16_t);
+            break;
+        }
+        case miopenIndexUint32: {
+            if(skip_many_configs_with_non_int8_index)
+            {
+                // test_pooling_test --all only test 5 uint32 cases
+                if(wsidx == 0)
+                {
+                    if(num_uint32_case > 5)
+                    {
+                        show_command();
+                        std::cout << "Warning: Config skipped for the default dataset to speed up "
+                                     "testing (wsidx == 0 && num_uint32_case > 5)"
+                                  << std::endl;
+                        return;
+                    }
+                    ++num_uint32_case;
+                }
+                else
+                {
+                    if(num_uint32_case_imgidx > 5)
+                    {
+                        show_command();
+                        std::cout << "Warning: Config skipped for the default dataset to speed up "
+                                     "testing (wsidx != 0 && num_uint32_case_imgidx > 5)"
+                                  << std::endl;
+                        return;
+                    }
+                    ++num_uint32_case_imgidx;
+                }
+            }
+            idx_sz = sizeof(uint32_t);
+            break;
+        }
+        case miopenIndexUint64: {
+            if(skip_many_configs_with_non_int8_index)
+            {
+                if(wsidx == 0)
+                {
+                    if(num_uint64_case > 5)
+                    {
+                        show_command();
+                        std::cout << "Warning: Config skipped for the default dataset to speed up "
+                                     "testing (wsidx == 0) && (num_uint64_case > 5)"
+                                  << std::endl;
+                        return;
+                    }
+                    ++num_uint64_case;
+                }
+                else
+                {
+                    if(num_uint64_case_imgidx > 5 && spt_dim == 2)
+                    {
+                        show_command();
+                        std::cout << "Warning: Config skipped to speed up testing of the "
+                                     "default dataset (wsidx != 0) && (num_uint64_case_imgidx > 5 "
+                                     "&& spt_dim == 2)"
+                                  << std::endl;
+                        return;
+                    }
+                    ++num_uint64_case_imgidx;
+                }
+            }
+            idx_sz = sizeof(uint64_t);
+            break;
+        }
+        }
+
+        auto input_desc = miopen::TensorDescriptor(this->type, in_shape);
+
+        if(spt_dim != 2 && spt_dim != 3)
+        {
+            show_command();
+            std::cout << "Warning: Config skipped becuse it is not supported " //
+                         "(spt_dim != 2 && spt_dim != 3)"
+                      << std::endl;
+            return;
+        }
+
+        for(int i = 0; i < spt_dim; i++)
+        {
+            if(lens[i] > (input_desc.GetLengths()[i + 2] + static_cast<uint64_t>(2) * pads[i]))
+            {
+                show_command();
+                std::cout << "Warning: Config skipped becuse it is invalid "
+                             "(lens[i] > (input_desc.GetLengths()[i + 2] + 2 * pads[i]))"
+                          << std::endl;
+                return;
+            }
+        }
+
+        if(full_set)
+        {
+            auto output_desc = filter.GetForwardOutputTensor(input_desc);
+            size_t total_mem =
+                3 * input_desc.GetNumBytes() + output_desc.GetNumBytes() +
+                idx_sz * output_desc.GetElementSize(); // estimate based on backward pass
+
+            size_t device_mem = get_handle().GetGlobalMemorySize();
+            if(total_mem >= device_mem)
+            {
+                show_command();
+                std::cout << "Config skipped because it requires " << total_mem
+                          << " Bytes to write all necessary tensors to GPU. GPU has " << device_mem
+                          << " Bytes of memory." << std::endl;
+                return;
+            }
+        }
+
+        std::vector<int> in_dim(input_desc.GetLengths().begin() + 2, input_desc.GetLengths().end());
+        std::vector<int> out_dim(spt_dim);
+        std::vector<int> ker_dim(filter.GetLengths().begin(), filter.GetLengths().end());
+#if TEST_PADDING_MODE == 1
+        if(filter.pmode == miopenPaddingSame)
+        {
+            if(std::any_of(filter.GetStrides().begin(), filter.GetStrides().end(), [](int i) {
+                   return i == 0;
+               }))
+                return;
+            for(int i = 0; i < spt_dim; i++)
+            {
+                filter.pads[i] =
+                    ((in_dim[i] % filter.GetStrides()[i] == 0)
+                         ? (std::max((ker_dim[i] - filter.GetStrides()[i]), 0))
+                         : (std::max((ker_dim[i] - (in_dim[i] % filter.GetStrides()[i])), 0))) /
+                    2;
+
+                out_dim[i] = std::ceil(static_cast<double>(in_dim[i]) / filter.strides[i]);
+            }
+
+            if(std::any_of(out_dim.begin(), out_dim.end(), [](int i) { return i <= 0; }))
+                return;
+        }
+        else if(filter.pmode == miopenPaddingValid)
+        {
+            if(std::any_of(filter.GetStrides().begin(), filter.GetStrides().end(), [](int i) {
+                   return i == 0;
+               }))
+                return;
+            for(int i = 0; i < spt_dim; i++)
+            {
+                filter.pads[i] = 0;
+
+                out_dim[i] = std::ceil(static_cast<double>(in_dim[i] - filter.lens[i] + 1) /
+                                       filter.strides[i]);
+            }
+
+            if(std::any_of(out_dim.begin(), out_dim.end(), [](int i) { return i <= 0; }))
+                return;
+        }
+#endif
+        switch(filter.GetIndexType())
+        {
+        case miopenIndexUint8: {
+            if(spt_dim == 3)
+            {
+                run_impl<uint8_t, 3>();
+            }
+            else
+            {
+                run_impl<uint8_t, 2>();
+            }
+            break;
+        }
+        case miopenIndexUint16: {
+            if(spt_dim == 3)
+            {
+                run_impl<uint16_t, 3>();
+            }
+            else
+            {
+                run_impl<uint16_t, 2>();
+            }
+            break;
+        }
+        case miopenIndexUint32: {
+            if(spt_dim == 3)
+            {
+                run_impl<uint32_t, 3>();
+            }
+            else
+            {
+                run_impl<uint32_t, 2>();
+            }
+            break;
+        }
+        case miopenIndexUint64: {
+            if(spt_dim == 3)
+            {
+                run_impl<uint64_t, 3>();
+            }
+            else
+            {
+                run_impl<uint64_t, 2>();
+            }
+            break;
+        }
+        }
     }
 };
